requirements.txt
# Core local server
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Spotify API
spotipy>=2.22.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
httpx>=0.25.0
requests>=2.31.0

# Apple data parsing (for iMessage binary data)
pytypedstream>=0.1.0

# Progress bars
tqdm>=4.65.0

# Database
sqlalchemy>=2.0.0

# Encryption (token-at-rest protection)
cryptography>=41.0.0

# Native window for bundled app
pywebview>=4.4.0

# Development/testing
pytest>=7.4.0
